@@ -186,13 +186,13 @@ FrameBuffer::FrameBuffer(GraphicsContext& context, Args&& args) : m_context(cont
         const ushort target_id = numbered_color_target.first;
         const auto& color_target = numbered_color_target.second;
 
-        if (notf::holds_alternative<RenderBufferPtr>(color_target)) {
-            if (const auto& color_buffer = notf::get<RenderBufferPtr>(color_target)) {
+        if (std::holds_alternative<RenderBufferPtr>(color_target)) {
+            if (const auto& color_buffer = std::get<RenderBufferPtr>(color_target)) {
                 notf_check_gl(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0 + target_id,
                                                         GL_RENDERBUFFER, color_buffer->get_id().value()));
             }
         }
-        else if (const auto& color_texture = notf::get<TexturePtr>(color_target)) {
+        else if (const auto& color_texture = std::get<TexturePtr>(color_target)) {
             notf_check_gl(glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0 + target_id,
                                                  color_texture->get_target(), color_texture->get_id().value(),
                                                  /* level = */ 0));
@@ -200,14 +200,14 @@ FrameBuffer::FrameBuffer(GraphicsContext& context, Args&& args) : m_context(cont
     }
 
     bool has_depth = false;
-    if (notf::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
-        if (const auto& depth_buffer = notf::get<RenderBufferPtr>(m_args.depth_target)) {
+    if (std::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
+        if (const auto& depth_buffer = std::get<RenderBufferPtr>(m_args.depth_target)) {
             notf_check_gl(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER,
                                                     depth_buffer->get_id().value()));
             has_depth = true;
         }
     }
-    else if (const auto& depth_texture = notf::get<TexturePtr>(m_args.depth_target)) {
+    else if (const auto& depth_texture = std::get<TexturePtr>(m_args.depth_target)) {
         notf_check_gl(glFramebufferTexture2D(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, depth_texture->get_target(),
                                              depth_texture->get_id().value(), /* level = */ 0));
         has_depth = true;
@@ -249,10 +249,10 @@ const TexturePtr& FrameBuffer::get_color_texture(const ushort id)
                             [id](const std::pair<ushort, ColorTarget>& target) -> bool { return target.first == id; });
     try {
         if (itr != std::end(m_args.color_targets)) {
-            return notf::get<TexturePtr>(itr->second);
+            return std::get<TexturePtr>(itr->second);
         }
     }
-    catch (notf::bad_variant_access&) {
+    catch (std::bad_variant_access&) {
         /* ignore */
     }
     NOTF_THROW(runtime_error, "FrameBuffer has no color attachment: {}", id);
@@ -269,15 +269,15 @@ void FrameBuffer::_deallocate()
     // deallocate all attached RenderBuffers
     for (const auto& numbered_color_target : m_args.color_targets) {
         const auto& color_target = numbered_color_target.second;
-        if (notf::holds_alternative<RenderBufferPtr>(color_target)) {
-            NOTF_ASSERT(notf::get<RenderBufferPtr>(color_target)); // RenderBuffer pointer should never be empty
-            render_buffer_access.deallocate(*notf::get<RenderBufferPtr>(color_target));
+        if (std::holds_alternative<RenderBufferPtr>(color_target)) {
+            NOTF_ASSERT(std::get<RenderBufferPtr>(color_target)); // RenderBuffer pointer should never be empty
+            render_buffer_access.deallocate(*std::get<RenderBufferPtr>(color_target));
         }
     }
 
-    if (notf::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
-        NOTF_ASSERT(notf::get<RenderBufferPtr>(m_args.depth_target)); // RenderBuffer pointer should never be empty
-        render_buffer_access.deallocate(*notf::get<RenderBufferPtr>(m_args.depth_target));
+    if (std::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
+        NOTF_ASSERT(std::get<RenderBufferPtr>(m_args.depth_target)); // RenderBuffer pointer should never be empty
+        render_buffer_access.deallocate(*std::get<RenderBufferPtr>(m_args.depth_target));
     }
 
     if (m_args.stencil_target) {
@@ -309,8 +309,8 @@ void FrameBuffer::_validate_arguments() const
         used_targets.insert(target_id);
 
         const auto& color_target = numbered_color_target.second;
-        if (notf::holds_alternative<RenderBufferPtr>(color_target)) {
-            if (const auto& color_buffer = notf::get<RenderBufferPtr>(color_target)) {
+        if (std::holds_alternative<RenderBufferPtr>(color_target)) {
+            if (const auto& color_buffer = std::get<RenderBufferPtr>(color_target)) {
                 if (color_buffer->get_type() != RenderBuffer::Type::COLOR) {
                     NOTF_THROW(runtime_error, "Cannot attach a RenderBuffer of type {} as color buffer",
                                type_to_str(color_buffer->get_type()));
@@ -329,8 +329,8 @@ void FrameBuffer::_validate_arguments() const
                 has_any_attachment = true;
             }
         }
-        else if (notf::holds_alternative<TexturePtr>(color_target)) {
-            if (notf::get<TexturePtr>(color_target)) {
+        else if (std::holds_alternative<TexturePtr>(color_target)) {
+            if (std::get<TexturePtr>(color_target)) {
                 has_any_attachment = true;
             }
         }
@@ -341,8 +341,8 @@ void FrameBuffer::_validate_arguments() const
 
     // depth attachment
     RenderBufferPtr depth_buffer;
-    if (notf::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
-        if ((depth_buffer = notf::get<RenderBufferPtr>(m_args.depth_target))) {
+    if (std::holds_alternative<RenderBufferPtr>(m_args.depth_target)) {
+        if ((depth_buffer = std::get<RenderBufferPtr>(m_args.depth_target))) {
             if (depth_buffer->get_type() != RenderBuffer::Type::DEPTH
                 && depth_buffer->get_type() != RenderBuffer::Type::DEPTH_STENCIL) {
                 NOTF_THROW(runtime_error, "Cannot attach a RenderBuffer of type {} as depth buffer",
@@ -362,8 +362,8 @@ void FrameBuffer::_validate_arguments() const
             has_any_attachment = true;
         }
     }
-    else if (notf::holds_alternative<TexturePtr>(m_args.depth_target)) {
-        if (notf::get<TexturePtr>(m_args.depth_target)) {
+    else if (std::holds_alternative<TexturePtr>(m_args.depth_target)) {
+        if (std::get<TexturePtr>(m_args.depth_target)) {
             has_any_attachment = true;
         }
     }
