@@ -1,9 +1,9 @@
 #pragma once
 
+#include <variant>
 #include <vector>
 
 #include "common/size2.hpp"
-#include "common/variant.hpp"
 #include "graphics/ids.hpp"
 
 NOTF_OPEN_NAMESPACE
@@ -135,10 +135,10 @@ public:
     using Access = access::_FrameBuffer<T>;
 
     /// A FrameBuffer's color target can be either a RenderBuffer or a Texture.
-    using ColorTarget = notf::variant<TexturePtr, RenderBufferPtr>;
+    using ColorTarget = std::variant<TexturePtr, RenderBufferPtr>;
 
     /// A FrameBuffer's depth target can be either a RenderBuffer or a Texture.
-    using DepthTarget = notf::variant<TexturePtr, RenderBufferPtr>;
+    using DepthTarget = std::variant<TexturePtr, RenderBufferPtr>;
 
     /// Arguments used to initialize a Texture.
     /// If you want to set both depth- and stencil targets, both have to refer to the same RenderBuffer and that
